tests/main_test.rs
use std::sync::Arc;
use mockall::predicate::*;
use mockall::mock;
use tokio::sync::Mutex;

mock! {
    pub Client {
        pub fn new(_: String) -> Self;
        pub fn chat_completion(&self, _: openai_api_rs::v1::chat_completion::ChatCompletionRequest) -> Result<openai_api_rs::v1::chat_completion::ChatCompletion, openai_api_rs::v1::error::OpenAIError>;
    }
}

mock! {
    pub ScreenCapture {
        pub fn capture_and_analyze_screen() -> Result<String, Box<dyn std::error::Error>>;
    }
}

mock! {
    pub SpeechRecognition {
        pub async fn listen_for_speech() -> Result<String, Box<dyn std::error::Error>>;
    listen_for_speech: ()}
}

mock! {
    pub TextToSpeech {
        pub fn speak(text: &str) -> Result<(), Box<dyn std::error::Error>>;
    speak: ()}
}

#[tokio::test]
async fn test_process_voice_input() {
    let mut mock_client = MockClient::new();
    mock_client
        .expect_chat_completion()
        .returning(|_| Ok(openai_api_rs::v1::chat_completion::ChatCompletionResponse {
            id: "test".to_string(),
            object: "chat.completion".to_string(),
            created: 0,
            model: "gpt-3.5-turbo".to_string(),
            choices: vec![openai_api_rs::v1::chat_completion::ChatCompletionChoice {
                index: 0,
                message: openai_api_rs::v1::chat_completion::ChatCompletionMessage {
                    role: openai_api_rs::v1::chat_completion::MessageRole::assistant,
                    content: Some("Test response".to_string()),
                    name: None,
                },
                finish_reason: Some("stop".to_string()),
                finish_details: None,
            }],
            usage: Some(openai_api_rs::v1::chat_completion::Usage {
                prompt_tokens: 0,
                completion_tokens: 0,
                total_tokens: 0,
            }),
            system_fingerprint: None,
            headers: None,
        }));

    let mut mock_speech_recognition = MockSpeechRecognition::new();
    mock_speech_recognition
        .expect_listen_for_speech()
        .returning(|| Ok("Test speech input".to_string()));

    let mut mock_tts = MockTextToSpeech::new();
    mock_tts
        .expect_speak()
        .with(eq("Test response"))
        .returning(|_| Ok(()));

    crate::listen_for_speech::listen_for_speech = mock_speech_recognition.listen_for_speech;
    crate::listen_for_speech::speak = mock_tts.speak;

    let result = crate::main::process_voice_input(&mock_client).await;
    assert!(result.is_ok());
}

#[tokio::test]
async fn test_process_screen_capture() {
    let mut mock_client = MockClient::new();
    mock_client
        .expect_chat_completion()
        .returning(|_| Ok(openai_api_rs::v1::chat_completion::ChatCompletion {
            id: "test".to_string(),
            object: "chat.completion".to_string(),
            created: 0,
            model: "gpt-3.5-turbo".to_string(),
            choices: vec![openai_api_rs::v1::chat_completion::ChatCompletionChoice {
                index: 0,
                message: openai_api_rs::v1::chat_completion::ChatCompletionMessage {
                    role: openai_api_rs::v1::chat_completion::MessageRole::assistant,
                    content: Some("Test analysis".to_string()),
                    name: None,
                },
                finish_reason: Some("stop".to_string()),
                finish_details: None,
            }],
            usage: Some(openai_api_rs::v1::chat_completion::Usage {
                prompt_tokens: 0,
                completion_tokens: 0,
                total_tokens: 0,
            }),
        }));

    let mut mock_screen_capture = MockScreenCapture::new();
    mock_screen_capture
        .expect_capture_and_analyze_screen()
        .returning(|| Ok("Test screen text".to_string()));

    crate::capture_and_analyse_screen::capture_and_analyze_screen = mock_screen_capture.capture_and_analyze_screen;

    let result = crate::main::process_screen_capture(&mock_client).await;
    assert!(result.is_ok());
}

#[test]
fn test_track_mouse_position() {
    let device_state = device_query::DeviceState::new();
    crate::main::track_mouse_position(&device_state);
}

#[tokio::test]
async fn test_main_loop() {
    let mock_client = Arc::new(Mutex::new(MockClient::new()));
    let mock_device_state = Arc::new(Mutex::new(device_query::DeviceState::new()));


    let mock_client_clone = Arc::clone(&mock_client);
    let mock_device_state_clone = Arc::clone(&mock_device_state);

    let handle = tokio::spawn(async move {
        // Run the main loop for a few iterations
        for _ in 0..3 {
            let client = mock_client_clone.lock().await;
            let device_state = mock_device_state_clone.lock().await;

            if let Err(e) = crate::main::process_voice_input(&client).await {
                eprintln!("Error in voice input processing: {}", e);
            }

            if let Err(e) = crate::main::process_screen_capture(&client).await {
                eprintln!("Error in screen capture processing: {}", e);
            }

            crate::main::track_mouse_position(&device_state);

            tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
        }
    });

    handle.await.unwrap();
}


#[tokio::test]
async fn test_listen_for_speech() {
}

#[test]
fn test_capture_and_analyze_screen() {
}

#[tokio::test]
async fn test_generate_response() {
    let mut mock_client = MockClient::new();
    mock_client
        .expect_chat_completion()
        .returning(|_| Ok(openai_api_rs::v1::chat_completion::ChatCompletion {
            id: "test".to_string(),
            object: "chat.completion".to_string(),
            created: 0,
            model: "gpt-3.5-turbo".to_string(),
            choices: vec![openai_api_rs::v1::chat_completion::ChatCompletionChoice {
                index: 0,
                message: openai_api_rs::v1::chat_completion::ChatCompletionMessage {
                    role: openai_api_rs::v1::chat_completion::MessageRole::assistant,
                    content: Some("Test response".to_string()),
                    name: None,
                },
                finish_reason: Some("stop".to_string()),
                finish_details: None,
            }],
            usage: Some(openai_api_rs::v1::chat_completion::Usage {
                prompt_tokens: 0,
                completion_tokens: 0,
                total_tokens: 0,
            }),
        }));

    let result = crate::ai::generate_response(&mock_client, "Test prompt");
    assert!(result.is_ok());
    assert_eq!(result.unwrap(), "Test response");
}

#[test]
fn test_speak() {
}